mettle/backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
aiosqlite==0.21.0
alembic==1.14.0

# Validation & Serialization
pydantic>=2.12.5
pydantic-settings==2.6.0
email-validator==2.2.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# CORS & Security
python-dotenv==1.0.1

# File handling
python-multipart==0.0.17
aiofiles==24.1.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.28.0

# Development
black==24.10.0
isort==5.13.0
greenlet==3.2.4
